+++ test-resources/infra/dequeueEvents/function.yaml
@@ -1,10 +1,10 @@
 AWSTemplateFormatVersion: "2010-09-09"
 
 Resources:
-  DequeueFunction:
+  DequeueEventsFunction:
     Type: AWS::Serverless::Function
     DependsOn:
-      - DequeueLogGroup
+      - DequeueEventsLogGroup
     Metadata:
       BuildMethod: esbuild
       BuildProperties:
@@ -12,11 +12,11 @@ Resources:
         Target: es2022
         Sourcemap: false
         EntryPoints:
-          - src/functions/dequeueEvents/dequeueHandler.ts
+          - src/functions/dequeueEvents/dequeueEventsHandler.ts
     Properties:
-      FunctionName: !Sub ${AWS::StackName}-dequeue
-      Handler: dequeueHandler.lambdaHandler
-      Role: !GetAtt DequeueRole.Arn
+      FunctionName: !Sub ${AWS::StackName}-dequeue-events
+      Handler: dequeueEventsHandler.lambdaHandler
+      Role: !GetAtt DequeueEventsRole.Arn
       Environment:
         Variables:
           TXMA_EVENT_TTL_DURATION_IN_SECONDS: 900
@@ -30,16 +30,16 @@ Resources:
         SecurityGroupIds:
           - !ImportValue devplatform-vpc-AWSServicesEndpointSecurityGroupId
 
-  DequeueLogGroup:
+  DequeueEventsLogGroup:
     Type: AWS::Logs::LogGroup
     Properties:
       RetentionInDays: 30
-      LogGroupName: !Sub /aws/lambda/${AWS::StackName}-dequeue
+      LogGroupName: !Sub /aws/lambda/${AWS::StackName}-dequeue-events
 
-  DequeueRole:
+  DequeueEventsRole:
     Type: AWS::IAM::Role
     Properties:
-      RoleName: !Sub ${AWS::StackName}-dequeue-role
+      RoleName: !Sub ${AWS::StackName}-dequeue-events-role
       AssumeRolePolicyDocument:
         Version: "2012-10-17"
         Statement:
@@ -101,13 +101,13 @@ Resources:
         - !Ref PermissionsBoundary
         - !Ref AWS::NoValue
 
-  DequeueEventSourceMapping:
+  DequeueEventsEventSourceMapping:
     Type: AWS::Lambda::EventSourceMapping
     Properties:
       BatchSize: 10
       Enabled: true
       EventSourceArn:
         Fn::ImportValue: !Sub ${BackendStackName}-txma-sqs-arn
-      FunctionName: !GetAtt DequeueFunction.Arn
+      FunctionName: !GetAtt DequeueEventsFunction.Arn
       FunctionResponseTypes:
         - ReportBatchItemFailures
