+++ test-resources/src/functions/dequeueEvents/dequeueEventsHandler.ts
